# Conflicts:
#	Makefile
#	core/pom.xml
#	core/src/main/java/feast/core/config/JobConfig.java
#	docs/getting-started/deploying-feast/docker-compose.md
#	docs/getting-started/deploying-feast/kubernetes.md
#	docs/reference/design-decisions.md
#	docs/user-guide/feature-retrieval.md
#	docs/user-guide/overview.md
#	docs/user-guide/statistics.md
#	infra/charts/feast/Chart.yaml
#	infra/charts/feast/README.md
#	infra/charts/feast/charts/feast-core/Chart.yaml
#	infra/charts/feast/charts/feast-core/README.md
#	infra/charts/feast/charts/feast-jupyter/Chart.yaml
#	infra/charts/feast/charts/feast-jupyter/README.md
#	infra/charts/feast/charts/feast-serving/Chart.yaml
#	infra/charts/feast/charts/feast-serving/README.md
#	infra/charts/feast/requirements.yaml
#	infra/docker-compose/docker-compose.batch.yml
#	infra/docker-compose/docker-compose.yml
#	infra/docker/jupyter/Dockerfile
#	infra/scripts/test-docker-compose.sh
#	ingestion/src/main/java/feast/ingestion/transform/ReadFromSource.java
#	ingestion/src/main/java/feast/ingestion/transform/fn/ValidateFeatureRowDoFn.java
#	ingestion/src/main/java/feast/ingestion/transform/specs/FilterRelevantFunction.java
#	pom.xml
#	protos/feast/core/Runner.proto
#	protos/feast/core/Store.proto
#	protos/feast/serving/ServingService.proto
#	sdk/python/feast/job.py
#	sdk/python/feast/loaders/file.py
#	sdk/python/feast/staging/storage_client.py
#	sdk/python/requirements-ci.txt
#	sdk/python/tests/loaders/test_file.py
#	sdk/python/tests/test_job.py
#	serving/pom.xml
#	serving/src/main/java/feast/serving/service/HistoricalServingService.java
#	serving/src/main/java/feast/serving/service/OnlineServingService.java
#	storage/api/src/main/java/feast/storage/api/retriever/HistoricalRetrievalResult.java
#	storage/api/src/main/java/feast/storage/api/retriever/HistoricalRetriever.java
#	storage/connectors/bigquery/src/main/java/feast/storage/connectors/bigquery/retriever/BigQueryHistoricalRetriever.java
#	storage/connectors/bigquery/src/test/java/feast/storage/connectors/bigquery/writer/BigQuerySinkTest.java
#	storage/connectors/redis/src/main/java/feast/storage/connectors/redis/retriever/RedisClusterOnlineRetriever.java
#	storage/connectors/redis/src/main/java/feast/storage/connectors/redis/retriever/RedisOnlineRetriever.java
#	storage/connectors/redis/src/main/java/feast/storage/connectors/redis/writer/RedisCustomIO.java
#	storage/connectors/redis/src/main/java/feast/storage/connectors/redis/writer/RedisFeatureSink.java

+++ sdk/python/feast/loaders/file.py
@@ -83,9 +83,7 @@ def export_source_to_staging_location(
             )
         else:
             # gs, s3 file provided as a source.
-            return get_staging_client(source_uri.scheme).list_files(
-                bucket=source_uri.hostname, path=source_uri.path
-            )
+            return get_staging_client(source_uri.scheme).list_files(source_uri)
     else:
         raise Exception(
             f"Only string and DataFrame types are allowed as a "
@@ -94,11 +92,16 @@ def export_source_to_staging_location(
 
     # Push data to required staging location
     get_staging_client(uri.scheme).upload_file(
-        source_path, uri.hostname, str(uri.path).strip("/") + "/" + file_name,
+        source_path, uri, str(uri.path).strip("/") + "/" + file_name,
     )
 
     # Clean up, remove local staging file
-    if dir_path and isinstance(source, pd.DataFrame) and len(dir_path) > 4:
+    if (
+        dir_path
+        and isinstance(source, pd.DataFrame)
+        and uri.scheme != "file"
+        and len(str(dir_path)) > 4
+    ):
         shutil.rmtree(dir_path)
 
     return [staging_location_uri.rstrip("/") + "/" + file_name]
